refactor project folder
