--- projeto/.env
@@ -1,7 +1,0 @@
-PYTHON_IMAGE_NAME=python-app-image
-MYSQL_IMAGE_NAME=review-db-image
-POSTGRES_NAME=imdb_reviews
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=admin
-POSTGRES_HOST=postgresql
-POSTGRES_PORT=5432
