--- projeto/database_scripts/requirements.txt
@@ -1,7 +1,0 @@
-selenium==4.10.0
-psycopg2-binary
-sqlalchemy==1.4.45
-tqdm==4.65.0
-torch==2.0.1
-transformers==4.30.2
-webdriver-manager==3.8.6
