--- projeto/docker-compose.yaml
@@ -1,29 +1,0 @@
-version: '3'
-
-services:
-  postgresql:
-    image: postgres:latest
-    volumes:
-      - postgres-vol:/var/lib/postgresql/data
-      - ./sql_scripts:/docker-entrypoint-initdb.d
-    environment:
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_USER: ${POSTGRES_USER}
-    ports:
-      - "5433:5432"
-  
-  database-scripts:
-    image: scraping-infer-image
-    build:
-      context: ./database_scripts
-    depends_on:
-      - postgresql
-    environment:
-      POSTGRES_NAME: ${POSTGRES_NAME}
-      POSTGRES_USER: ${POSTGRES_USER}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_HOST: ${POSTGRES_HOST}
-      POSTGRES_PORT: ${POSTGRES_PORT}
-
-volumes:
-  postgres-vol:
