--- projeto/requirements.txt
@@ -1,4 +1,0 @@
-selenium==4.9.1
-mysql-connector-python
-psycopg2==2.9.5
-sqlalchemy==1.4.45
